+++ packages/hooks/useLocalStorage.ts
@@ -1,4 +1,4 @@
-import { ref, Ref, watch, computed, ComputedRef } from "vue";
+import { ref, watch, computed, ComputedRef } from "vue";
 
 /**
  * 设置存储值的函数类型